# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  pending: 'pending',
  paid: 'paid',
  expired: 'expired',
} as const;

export interface Invoice {
  id: string;
  invoiceNumber: string;
  customer: string;
  /** @nullable */
  note?: string | null;
  amount: number;
  currency: string;
  cryptoAmount: number;
  cryptoSymbol: string;
  status: InvoiceStatus;
  createdAt: string;
  expiresAt: string;
  network: string;
  fee: number;
  address?: string;
}

export type InvoiceInputCurrency = typeof InvoiceInputCurrency[keyof typeof InvoiceInputCurrency];


export const InvoiceInputCurrency = {
  USD: 'USD',
  EUR: 'EUR',
  THB: 'THB',
} as const;

export type InvoiceInputCryptoSymbol = typeof InvoiceInputCryptoSymbol[keyof typeof InvoiceInputCryptoSymbol];


export const InvoiceInputCryptoSymbol = {
  USDT: 'USDT',
} as const;

export type InvoiceInputNetwork = typeof InvoiceInputNetwork[keyof typeof InvoiceInputNetwork];


export const InvoiceInputNetwork = {
  TRC20: 'TRC20',
  ERC20: 'ERC20',
  BEP20: 'BEP20',
} as const;

export interface InvoiceInput {
  /** @minLength 1 */
  customer: string;
  note?: string;
  /** @minimum 0.01 */
  amount: number;
  currency: InvoiceInputCurrency;
  cryptoSymbol: InvoiceInputCryptoSymbol;
  network: InvoiceInputNetwork;
  /**
     * @minimum 15
     * @maximum 1440
     */
  expiresInMinutes?: number;
}

export interface CryptoPrice {
  symbol: string;
  name: string;
  price: number;
  change24h: number;
  updatedAt: string;
}

export interface Dashboard {
  totalReceived: number;
  pendingAmount: number;
  successRate: number;
  activeInvoices: number;
  recentActivity: Invoice[];
}

export interface Settings {
  botName: string;
  botConnected: boolean;
  serviceFeePercent: number;
  enabledCurrencies: string[];
  defaultExpiryMinutes: number;
}

export type SettingsUpdateEnabledCurrenciesItem = typeof SettingsUpdateEnabledCurrenciesItem[keyof typeof SettingsUpdateEnabledCurrenciesItem];


export const SettingsUpdateEnabledCurrenciesItem = {
  USDT: 'USDT',
} as const;

export interface SettingsUpdate {
  /** @minLength 1 */
  botName?: string;
  /**
     * @minimum 0
     * @maximum 20
     */
  serviceFeePercent?: number;
  enabledCurrencies?: SettingsUpdateEnabledCurrenciesItem[];
  /**
     * @minimum 15
     * @maximum 1440
     */
  defaultExpiryMinutes?: number;
}

export type ListInvoicesParams = {
status?: ListInvoicesStatus;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListInvoicesStatus = typeof ListInvoicesStatus[keyof typeof ListInvoicesStatus];


export const ListInvoicesStatus = {
  all: 'all',
  pending: 'pending',
  paid: 'paid',
  expired: 'expired',
} as const;